Rename interface files for consistency
