+++ backend/graphql/resolvers/entityResolvers.ts
@@ -12,7 +12,7 @@ import FileStorageService from "../../services/implementations/fileStorageServic
 import {
   EntityRequestDTO,
   EntityResponseDTO,
-} from "../../services/interfaces/IEntityService";
+} from "../../services/interfaces/entityService";
 import { generateCSV } from "../../utilities/CSVUtils";
 
 const defaultBucket = process.env.DEFAULT_BUCKET || "";
