+++ backend/services/interfaces/entityService.ts
