+++ backend/services/interfaces/postingService.ts
